+++ util.h
@@ -1,5 +1,5 @@
 #include <stdio.h>
 
-void print_array(int[], size_t);
+void print_array(int[], int);
 void print_sort_header(const char *);
-void print_sort(int[], size_t, void (*sort)(int[], size_t));
+void print_sort(int[], int, void (*sort)(int[], int));
